Here, the Python code for spatial analysis.
This is like a Bootcamp which I did on Udemy. I learned how to do spatial analysis with Covid-19 data, how to make a choropleth map, and how to filter data for a choropleth map. I also had to apply many Pandas functions to better filter and screen the data for analysis.